+++ docs/RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 ## 5.0.0
 
 ### New
+* acorn: Trim link title of electronic resources and always show popup for links
 * cwmars: Add "More information" button to Item Details screen, which links to an expanded
   description with series details and read-alike recommendations
 * Update minimum Android version to 6.0 (API 23)