+++ docs/RELEASE_NOTES.md
@@ -1,11 +1,13 @@
 ## 5.0.0
 
 ### New
+* acorn: Trim link title of electronic resources and always show popup for links
 * cwmars: Add "More information" button to Item Details screen, which links to an expanded
   description with series details and read-alike recommendations
 * Update minimum Android version to 6.0 (API 23)
 
 ### Fixed
+* Fixed bug where copy summary for holdable items with electronic resources would display a hostname instead of the number of available copies
 * Made app download smaller by enabling R8 full mode optimization
 
 ### Internal